Moved the url creator into it's own file
Fixed/updated tests, and got it all working
